Added Javadocs to most files and changed access modifiers to private or package-private.
ChatroomGUI.java:
	Removed redundant check for empty messages.
ReceiveMessageThread.java:
	Made methods private/package-private.
	Added Javadocs.
Salt.java:
	storePassword now returns a boolean indicating whether it succeeded or not.
	Added Javadocs.
SendMessageThread.java:
	Made methods package-private.
	Added Javadocs.
SocketServer.java:
	Condensed everything in the main while loop into a single try block.
	Added Javadocs.
SocketServerThread.java:
	Made methods package-private.
	Added Javadocs.

+++ src/GUI/Salt.java
@@ -18,6 +18,10 @@ public class Salt {
         this.file = file;
     }
 
+    /**
+     * Generates a salt to be used for password storage.
+     * @return A salt String to be saved to file for use in password authentication.
+     */
     private String generateSalt(){
         try {
             byte[] salt = SecureRandom.getInstance(SALT_ALGORITHM).generateSeed(SALT_LENGTH);
@@ -28,6 +32,11 @@ public class Salt {
         return null;
     }
 
+    /**
+     * Hashes a salted password to save to file for use in password authentication.
+     * @param saltedPassword A String from the concatenation of the salt String and password String
+     * @return A hash String to be saved to file for use in password authentication.
+     */
     private String generateHash(String saltedPassword){
         try {
             MessageDigest md = MessageDigest.getInstance(HASH_ALGORITHM);
@@ -38,20 +47,47 @@ public class Salt {
         }catch (NoSuchAlgorithmException nsa){
             //This should also never happen.
         }
-        return "";
+        return null;
     }
 
-    void storePassword(String username, String password){
-        try {
-            PrintWriter fileOut = new PrintWriter(new FileWriter(file, true));
-            String salt = generateSalt();
-            fileOut.print(username + " " + salt + " " + generateHash(salt + password) + "\n");
-            fileOut.close();
-        } catch (IOException e){
-            System.out.println("Error writing hashed-password to file.");
+    /**
+     * Takes in a username and password, and attempts to create a salt and hash
+     * to store login information. Checks if the username already exists first.
+     * Writes the username, salt, and hash to a text file.
+     * @param username String containing the username to save.
+     * @param password String containing the password to salt and hash.
+     * @return true if the information is successfully saved to file. false if the username is
+     * already contained in the file, or an exception occurs.
+     */
+    boolean storePassword(String username, String password){
+        //Without this, one user may have multiple salts and hashes saved to file.
+        //However, as-is, a password cannot be changed once it is set.
+        if(usernameAlreadyInDatabase(username)){
+            System.out.println("Username is already in database.");
+            return false;
+        }
+        else {
+            try {
+                PrintWriter fileOut = new PrintWriter(new FileWriter(file, true));
+                String salt = generateSalt();
+                fileOut.print(username + " " + salt + " " + generateHash(salt + password) + "\n");
+                fileOut.close();
+                return true;
+            } catch (IOException e) {
+                System.out.println("Error writing hashed-password to file.");
+                return false;
+            }
         }
     }
 
+    /**
+     * Takes in a username and password, and tries to see if the hash of the entered
+     * password matches the hash saved to file.
+     * @param username a String containing the username to look for
+     * @param password a String containing the password
+     * @return true if the entered password hashes to the right value. false if it doesn't, if the
+     * entered username does not exist in the file, or an exception was raised reading from the file.
+     */
     boolean authenticatePassword(String username, String password){
         try {
             BufferedReader fileIn = new BufferedReader(new FileReader(this.file));
@@ -61,19 +97,42 @@ public class Salt {
                 //Do nothing until you reach EOF or the line containing the username.
             }
 
-            if(line == null)
+            if(line == null) {
                 System.out.println("Username not in database.");
+                fileIn.close();
+                return false;
+            }
 
             //Username was found
             else{
                 String[] fields = line.split(" ");
                 String userGeneratedHash = generateHash(fields[1] + password);
+                fileIn.close();
                 return userGeneratedHash.equals(fields[2]);
             }
-            fileIn.close();
         } catch(IOException e){
             System.out.println("Error reading file.");
+            return false;
         }
-        return false;
+    }
+
+    /**
+     * Goes through a text file to look for a username.
+     * @param username a String of the username to look for
+     * @return true if the username is contained in the file.
+     */
+    private boolean usernameAlreadyInDatabase(String username){
+        String line = null;
+        try {
+            BufferedReader fileIn = new BufferedReader(new FileReader(this.file));
+
+            while (((line = fileIn.readLine()) != null) && (!line.split(" ")[0].equals(username))) {
+                //Do nothing until you reach EOF or the line containing the username.
+            }
+        } catch (IOException e){
+            System.out.println("Error reading file.");
+        }
+        //If line is still null, there was an exception, or EOF was reached and username was not found.
+        return line != null;
     }
 }
